requirements.txt
discord.py>=2.3.2
anthropic>=0.50.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
dateparser>=1.2.0
